+++ Pod/Classes/CollieGalleryDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  CollieGalleryDelegate.swift
+//  Pods
+//
+//  Created by Guilherme Munhoz on 5/11/16.
+//
+//
+
+import UIKit
+
+@objc public protocol CollieGalleryDelegate: class {
+    optional func gallery(gallery: CollieGallery, indexChangedTo index: Int)
+    
+}